Fix indentation of record fields, mostly
I noticed that writing records was broken. That is, instead of (| is point):

type ARecord =
    { FieldOne: int
      |}

One always got:

type ARecord =
    { FieldOne: int
  }

Which is not syntactically correct. I added a test for it, and found the problem
was fsharp-outdent-re, which had `}` as a character. Ideal!

There are good odds this affects other parenthetical expressions; will need to
get to those eventually.